+++ src/components/common/register.vue
@@ -1,0 +1,5 @@
+<template>
+  <button class="button green" disabled>
+    <strong>{{ $t('components.register.register_text') }}</strong>
+  </button>
+</template>